\name{getPvalBH}
\Rdversion{1.1}
\alias{getPvalBH}

\title{
 Corrected p values using Benjamini & Hochberg approach
}
\description{
 To be suypplied
}

\usage{
getPvalBH(x, map)
}


\arguments{
  \item{x}{a list containing p values}
  \item{map}{genomic information for each blok of CNVs}
}

\keyword{misc}

