\name{multiCNVassoc}
\alias{multiCNVassoc}
\alias{plot.multiCNVassoc}

\title{
Association between several CNVs and disease
}
\description{
To be supplied
}

\usage{
multiCNVassoc(x, formula, ...)

\method{plot}{multiCNVassoc}(x, map, ...)

}

\arguments{
  \item{x}{TBD}
  \item{formula}{TBD}
  \item{map}{TBD}
  \item{\dots}{TBD}
}

\keyword{misc}

